refactor: move sqlite metadata and query execution
