+++ src/infra/adapters/sqlite/sqlite3/mod.rs
@@ -1,7 +1,6 @@
 mod error;
 mod executor;
+mod metadata;
 pub(super) mod parser;
 
-#[cfg(test)]
-pub(in crate::adapters::sqlite) use executor::BUSY_TIMEOUT_MS;
 pub(super) use executor::SqliteCli;
